Enables mergin resources from the travel-agent.yml configuration
more generic pipelines as resources specific to a product deployment can be set in the travel-agent.yml config and not in the manifest.ego.

Concourse allows inputs for tasks to be optional (https://concourse-ci.org/tasks.html#input-optional). In this way you can have a task that allows inputs A, B and C, but the travel-agent.yml config only pulls a subset of them.