Adds EnvsWithFeature and HasEnvsWithFeature
Allows users to globally check if a certain feature is being use.
This is useful when adding groups to a pipeline manifest template. Some features only apply to certain environments and to check if it worth adding a group for those jobs you currently need to iterate through all envs.